requirements.txt
gradio
openai
anthropic
python-dotenv